[core] subdivide the types module
